Translation interface fixes (see desc for info)
I changed the way that the language is stored in the session. Rather
than storing the entire language file into the session, I've replaced it
with just the language name, and make the server figure it out instead
of us filling up our sessiondb quicker because of the whole language
file being added.

After this, it created a weird edge case with the name of the language,
because it would be modified whenever someone accessed / (for
personalization purposes), so this was fixed by changing the "name" key
to "_name" in each language file, this way it won't conflict, and we can
have a well-running system.